+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.0]
+### Changed
+- fetcher passed to ExporterMeta https://github.com/ffeast/finam-export/pull/22
+
 ## [4.0.2]
 ### Fixed
 - TICKS data download in some cases might miss some days (https://github.com/ffeast/finam-export/issues/16)